+++ src/com/kmne68/main/Trail.java
@@ -23,11 +23,11 @@ public class Trail extends GameObject {
 	private Handler handler;
 	private Color color;
 	
-	private int width;
-	private int height;
+	private float width;
+	private float height;
 
 	
-	public Trail(int x, int y, ID id, Color color, int width, int height, float life, Handler handler) {
+	public Trail(float x, float y, ID id, Color color, float width, float height, float life, Handler handler) {
 		super(x, y, id);
 		
 		this.color = color;
@@ -55,7 +55,7 @@ public class Trail extends GameObject {
 		
 		g2d.setComposite(makeTransparent(alpha));
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect( (int) x, (int) y, (int) width, (int) height);
 		g2d.setComposite(makeTransparent(1));
 		
 	}
